pyproject.toml
[tool.poetry]
name = "sistema-pos"
version = "0.1.0"
description = "Sistema punto de venta"
authors = ["tomiban <tbanchio15@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
flet = "0.24.*"
sqlalchemy = "*"
alembic = "*"
mysql-connector-python = "*"
python-dotenv = "*"
fpdf = "*"
pandas = "*"
flet-easy = {extras = ["all"], version = "^0.2.7"}
werkzeug = "^3.1.3"


[tool.poetry.group.dev.dependencies]
pytest = "*"
black = "*"
flake8 = "*"
isort = "*"
pre-commit = "*"
commitizen = "*"
poethepoet = "^0.31.1"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 150
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
force_single_line = true

[tool.black]
line-length = 150
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/foo.py  # exclude a file named foo.py in the root of the project
'''

[tool.flake8]
max-line-length = 150
exclude = [".git", "__pycache__", "build", "dist"]

[tool.poe.tasks]
dev = { cmd = "python app/main.py", env = { "ENV_NAME" = "development" } }
prod = { cmd = "python app/main.py", env = { "ENV_NAME" = "production" } }

format = { sequence = ["_black", "_isort"] }
_black = { cmd = "black ." }
_isort = { cmd = "isort ." }

lint = { sequence = ["_flake8"] }
_flake8 = { cmd = "flake8" }

check = { sequence = ["format", "lint"] }
test = { cmd = "pytest" }
test-cov = { cmd = "pytest --cov=sistema_pos --cov-report=term-missing" }
db-revision = { cmd = "alembic revision --autogenerate -m", help = "Create new alembic revision. Usage: poe revision 'message'"}
db-migrate = { cmd = "alembic upgrade head" }
db-rollback = { cmd = "alembic downgrade -1" }

clean = { shell = "rm -rf .pytest_cache .coverage htmlcov __pycache__ dist build" }
precommit = { cmd = "pre-commit run --all-files" }
commit = { cmd = "cz commit" }
version = { cmd = "cz bump" }
